+++ proto/control-api/src/control/endpoint/mod.rs
@@ -4,6 +4,7 @@ pub mod web_rtc_play;
 pub mod web_rtc_publish;
 
 use derive_more::{Display, From, Into};
+use ref_cast::RefCast;
 
 #[doc(inline)]
 pub use self::{web_rtc_play::WebRtcPlay, web_rtc_publish::WebRtcPublish};
@@ -27,10 +28,10 @@ pub enum Endpoint {
 impl Endpoint {
     /// Returns [`Id`] of this [`Endpoint`].
     #[must_use]
-    pub fn id(&self) -> Id {
+    pub fn id(&self) -> &Id {
         match self {
-            Endpoint::WebRtcPublish(publish) => publish.id.clone().into(),
-            Endpoint::WebRtcPlay(play) => play.id.clone().into(),
+            Self::WebRtcPublish(publish) => Id::ref_cast(publish.id.as_ref()),
+            Self::WebRtcPlay(play) => Id::ref_cast(play.id.as_ref()),
         }
     }
 }
@@ -39,10 +40,21 @@ impl Endpoint {
 ///
 /// [`Element`]: crate::Element
 #[derive(
-    Clone, Debug, Display, Eq, From, Hash, Into, Ord, PartialEq, PartialOrd,
+    Clone,
+    Debug,
+    Display,
+    Eq,
+    From,
+    Hash,
+    Into,
+    Ord,
+    PartialEq,
+    PartialOrd,
+    RefCast,
 )]
 #[from(types(String, web_rtc_publish::Id, web_rtc_play::Id))]
 #[into(owned(types(String)))]
+#[repr(transparent)]
 pub struct Id(Box<str>);
 
 // TODO: Derive via `derive::From` once it's capable to.