+++ proto/control-api/src/grpc/convert/api.rs
@@ -1,7 +1,8 @@
-//! Definitions of conversions from [`ControlApi`] spec into generated from
-//! `api.proto` spec via [`tonic-build`] and vise-versa.
+//! Conversions between [`ControlApi`] types and the ones generated from
+//! `api.proto` [gRPC] spec.
 //!
 //! [`ControlApi`]: crate::ControlApi
+//! [gRPC]: https://grpc.io
 
 use std::time::Duration;
 
@@ -11,20 +12,18 @@ use crate::{
         web_rtc_publish::{AudioSettings, P2pMode, Policy, VideoSettings},
         WebRtcPlay, WebRtcPublish,
     },
-    grpc::{
-        api as proto, conversions::TryFromProtobufError, CallbackUrlParseError,
-    },
+    grpc::{api as proto, convert::ProtobufError, CallbackUrlParseError},
     member::{self, Credentials},
     Element, Endpoint, Fid, Member, Ping, Pong, Room,
 };
 
 impl TryFrom<proto::CreateRequest> for Request {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(value: proto::CreateRequest) -> Result<Self, Self::Error> {
-        let parent_fid = value.parent_fid;
-        let el = value.el.ok_or_else(|| {
-            TryFromProtobufError::EmptyElementId(parent_fid.clone())
+    fn try_from(req: proto::CreateRequest) -> Result<Self, Self::Error> {
+        let parent_fid = req.parent_fid;
+        let el = req.el.ok_or_else(|| {
+            ProtobufError::NoElementForId(parent_fid.as_str().into())
         })?;
 
         if parent_fid.is_empty() {
@@ -45,7 +44,7 @@ impl TryFrom<proto::CreateRequest> for Request {
                     endpoint,
                 })
             }
-            Fid::Endpoint { .. } => Err(TryFromProtobufError::ParseFidError(
+            Fid::Endpoint { .. } => Err(ProtobufError::ParseFidErr(
                 ParseFidError::TooManyPaths(parent_fid.into()),
             )),
         }
@@ -53,12 +52,12 @@ impl TryFrom<proto::CreateRequest> for Request {
 }
 
 impl TryFrom<proto::ApplyRequest> for Request {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(value: proto::ApplyRequest) -> Result<Self, Self::Error> {
-        let parent_fid = value.parent_fid;
-        let el = value.el.ok_or_else(|| {
-            TryFromProtobufError::EmptyElementId(parent_fid.clone())
+    fn try_from(req: proto::ApplyRequest) -> Result<Self, Self::Error> {
+        let parent_fid = req.parent_fid;
+        let el = req.el.ok_or_else(|| {
+            ProtobufError::NoElementForId(parent_fid.as_str().into())
         })?;
 
         match parent_fid.parse::<Fid>()? {
@@ -69,9 +68,7 @@ impl TryFrom<proto::ApplyRequest> for Request {
                     member: Box::new(member),
                 })
             }
-            Fid::Endpoint { .. } => {
-                Err(TryFromProtobufError::UnimplementedCall)
-            }
+            Fid::Endpoint { .. } => Err(ProtobufError::Unimplemented),
         }
     }
 }
@@ -80,10 +77,9 @@ impl From<Request> for proto::CreateRequest {
     fn from(req: Request) -> Self {
         let (parent_fid, el) = match req {
             Request::Room(room) => (String::new(), room.into()),
-            Request::Member { room_id, member } => (
-                Fid::Room { id: room_id }.to_string(),
-                member.as_ref().clone().into(),
-            ),
+            Request::Member { room_id, member } => {
+                (Fid::Room { id: room_id }.to_string(), (*member).into())
+            }
             Request::Endpoint {
                 room_id,
                 member_id,
@@ -127,7 +123,7 @@ impl From<Request> for proto::ApplyRequest {
                 endpoint,
             } => (
                 Fid::Endpoint {
-                    id: endpoint.id(),
+                    id: endpoint.id().clone(),
                     room_id,
                     member_id,
                 },
@@ -144,19 +140,17 @@ impl From<Request> for proto::ApplyRequest {
 
 impl From<Element> for proto::Element {
     fn from(el: Element) -> Self {
-        use proto::element::El as proto_el;
+        use proto::element::El;
 
         Self {
             el: Some(match el {
-                Element::Member(member) => {
-                    proto_el::Member(member.as_ref().clone().into())
-                }
-                Element::Room(room) => proto_el::Room(room.into()),
+                Element::Member(member) => El::Member((*member).into()),
+                Element::Room(room) => El::Room(room.into()),
                 Element::Endpoint(Endpoint::WebRtcPlay(play)) => {
-                    proto_el::WebrtcPlay(play.into())
+                    El::WebrtcPlay(play.into())
                 }
                 Element::Endpoint(Endpoint::WebRtcPublish(publish)) => {
-                    proto_el::WebrtcPub(publish.into())
+                    El::WebrtcPub(publish.into())
                 }
             }),
         }
@@ -164,20 +158,18 @@ impl From<Element> for proto::Element {
 }
 
 impl TryFrom<proto::Element> for Element {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
     fn try_from(el: proto::Element) -> Result<Self, Self::Error> {
-        use proto::element::El as proto_el;
+        use proto::element::El;
 
-        Ok(match el.el.ok_or(TryFromProtobufError::EmptyElement)? {
-            proto_el::Member(member) => {
-                Self::Member(Box::new(member.try_into()?))
-            }
-            proto_el::Room(room) => Self::Room(room.try_into()?),
-            proto_el::WebrtcPlay(play) => {
+        Ok(match el.el.ok_or(ProtobufError::NoElement)? {
+            El::Member(member) => Self::Member(Box::new(member.try_into()?)),
+            El::Room(room) => Self::Room(room.try_into()?),
+            El::WebrtcPlay(play) => {
                 Self::Endpoint(Endpoint::WebRtcPlay(play.try_into()?))
             }
-            proto_el::WebrtcPub(publish) => {
+            El::WebrtcPub(publish) => {
                 Self::Endpoint(Endpoint::WebRtcPublish(publish.into()))
             }
         })
@@ -188,7 +180,7 @@ impl From<Element> for proto::create_request::El {
     fn from(el: Element) -> Self {
         match el {
             Element::Room(room) => room.into(),
-            Element::Member(member) => member.as_ref().clone().into(),
+            Element::Member(member) => (*member).into(),
             Element::Endpoint(endpoint) => endpoint.into(),
         }
     }
@@ -198,52 +190,48 @@ impl From<Element> for proto::apply_request::El {
     fn from(el: Element) -> Self {
         match el {
             Element::Room(room) => room.into(),
-            Element::Member(member) => member.as_ref().clone().into(),
+            Element::Member(member) => (*member).into(),
             Element::Endpoint(endpoint) => endpoint.into(),
         }
     }
 }
 
 impl TryFrom<proto::create_request::El> for Room {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(proto: proto::create_request::El) -> Result<Self, Self::Error> {
-        use proto::create_request::El as proto_el;
+    fn try_from(val: proto::create_request::El) -> Result<Self, Self::Error> {
+        use proto::create_request::El;
 
-        match proto {
-            proto_el::Room(room) => room.try_into(),
-            proto_el::Member(proto::Member { id, .. })
-            | proto_el::WebrtcPub(proto::WebRtcPublishEndpoint {
-                id, ..
-            })
-            | proto_el::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
-                Err(TryFromProtobufError::ExpectedOtherElement("Room", id))
+        match val {
+            El::Room(room) => room.try_into(),
+            El::Member(proto::Member { id, .. })
+            | El::WebrtcPub(proto::WebRtcPublishEndpoint { id, .. })
+            | El::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
+                Err(ProtobufError::ExpectedElement("Room", id.into()))
             }
         }
     }
 }
 
 impl TryFrom<proto::apply_request::El> for Room {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(proto: proto::apply_request::El) -> Result<Self, Self::Error> {
-        use proto::apply_request::El as proto_el;
+    fn try_from(val: proto::apply_request::El) -> Result<Self, Self::Error> {
+        use proto::apply_request::El;
 
-        match proto {
-            proto_el::Room(room) => room.try_into(),
-            proto_el::Member(proto::Member { id, .. })
-            | proto_el::WebrtcPub(proto::WebRtcPublishEndpoint {
-                id, ..
-            })
-            | proto_el::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
-                Err(TryFromProtobufError::ExpectedOtherElement("Room", id))
+        match val {
+            El::Room(room) => room.try_into(),
+            El::Member(proto::Member { id, .. })
+            | El::WebrtcPub(proto::WebRtcPublishEndpoint { id, .. })
+            | El::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
+                Err(ProtobufError::ExpectedElement("Room", id.into()))
             }
         }
     }
 }
 
 impl TryFrom<proto::Room> for Room {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
     fn try_from(room: proto::Room) -> Result<Self, Self::Error> {
         Ok(Self {
@@ -253,7 +241,7 @@ impl TryFrom<proto::Room> for Room {
                 .into_iter()
                 .map(|(id, room_element)| {
                     room_element.el.map_or(
-                        Err(TryFromProtobufError::EmptyElementId(id)),
+                        Err(ProtobufError::NoElementForId(id.into())),
                         |el| {
                             Member::try_from(el)
                                 .map(|member| (member.id.clone(), member))
@@ -291,75 +279,69 @@ impl From<Room> for proto::Room {
 }
 
 impl TryFrom<proto::create_request::El> for Member {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(proto: proto::create_request::El) -> Result<Self, Self::Error> {
-        use proto::create_request::El as proto_el;
+    fn try_from(val: proto::create_request::El) -> Result<Self, Self::Error> {
+        use proto::create_request::El;
 
-        match proto {
-            proto_el::Member(member) => member.try_into(),
-            proto_el::Room(proto::Room { id, .. })
-            | proto_el::WebrtcPub(proto::WebRtcPublishEndpoint {
-                id, ..
-            })
-            | proto_el::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
-                Err(TryFromProtobufError::ExpectedOtherElement("Member", id))
+        match val {
+            El::Member(member) => member.try_into(),
+            El::Room(proto::Room { id, .. })
+            | El::WebrtcPub(proto::WebRtcPublishEndpoint { id, .. })
+            | El::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
+                Err(ProtobufError::ExpectedElement("Member", id.into()))
             }
         }
     }
 }
 
 impl TryFrom<proto::apply_request::El> for Member {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(proto: proto::apply_request::El) -> Result<Self, Self::Error> {
-        use proto::apply_request::El as proto_el;
+    fn try_from(val: proto::apply_request::El) -> Result<Self, Self::Error> {
+        use proto::apply_request::El;
 
-        match proto {
-            proto_el::Member(member) => member.try_into(),
-            proto_el::Room(proto::Room { id, .. })
-            | proto_el::WebrtcPub(proto::WebRtcPublishEndpoint {
-                id, ..
-            })
-            | proto_el::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
-                Err(TryFromProtobufError::ExpectedOtherElement("Member", id))
+        match val {
+            El::Member(member) => member.try_into(),
+            El::Room(proto::Room { id, .. })
+            | El::WebrtcPub(proto::WebRtcPublishEndpoint { id, .. })
+            | El::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
+                Err(ProtobufError::ExpectedElement("Member", id.into()))
             }
         }
     }
 }
 
 impl TryFrom<proto::room::element::El> for Member {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(proto: proto::room::element::El) -> Result<Self, Self::Error> {
-        use proto::room::element::El as proto_el;
+    fn try_from(val: proto::room::element::El) -> Result<Self, Self::Error> {
+        use proto::room::element::El;
 
-        match proto {
-            proto_el::Member(member) => Self::try_from(member),
-            proto_el::WebrtcPub(proto::WebRtcPublishEndpoint {
-                id, ..
-            })
-            | proto_el::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
-                Err(TryFromProtobufError::ExpectedOtherElement("Member", id))
+        match val {
+            El::Member(member) => Self::try_from(member),
+            El::WebrtcPub(proto::WebRtcPublishEndpoint { id, .. })
+            | El::WebrtcPlay(proto::WebRtcPlayEndpoint { id, .. }) => {
+                Err(ProtobufError::ExpectedElement("Member", id.into()))
             }
         }
     }
 }
 
 impl TryFrom<proto::Member> for Member {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
     fn try_from(member: proto::Member) -> Result<Self, Self::Error> {
-        /// Tries to parse [`Duration`] and maps error to
+        /// Tries to parse [`Duration`] and maps its error into a
         /// [`TryFromProtobufError`].
         fn parse_duration<T: TryInto<Duration>>(
             duration: Option<T>,
             member_id: &str,
             field: &'static str,
-        ) -> Result<Option<Duration>, TryFromProtobufError> {
+        ) -> Result<Option<Duration>, ProtobufError> {
             #[allow(clippy::map_err_ignore)]
             duration.map(TryInto::try_into).transpose().map_err(|_| {
-                TryFromProtobufError::NegativeDuration(member_id.into(), field)
+                ProtobufError::InvalidDuration(member_id.into(), field)
             })
         }
 
@@ -380,10 +362,10 @@ impl TryFrom<proto::Member> for Member {
                 .into_iter()
                 .map(|(id, member_el)| {
                     member_el.el.map_or(
-                        Err(TryFromProtobufError::EmptyElementId(id)),
+                        Err(ProtobufError::NoElementForId(id.into())),
                         |el| {
                             Endpoint::try_from(el)
-                                .map(|endpoint| (endpoint.id(), endpoint))
+                                .map(|endpnt| (endpnt.id().clone(), endpnt))
                         },
                     )
                 })
@@ -450,12 +432,12 @@ impl From<Member> for proto::room::Element {
 }
 
 impl From<proto::member::Credentials> for Credentials {
-    fn from(from: proto::member::Credentials) -> Self {
-        use proto::member::Credentials as C;
+    fn from(val: proto::member::Credentials) -> Self {
+        use proto::member::Credentials as Creds;
 
-        match from {
-            C::Hash(hash) => Self::Hash(hash.into()),
-            C::Plain(plain) => Self::Plain(plain.into()),
+        match val {
+            Creds::Hash(hash) => Self::Hash(hash.into()),
+            Creds::Plain(plain) => Self::Plain(plain.into()),
         }
     }
 }
@@ -470,33 +452,29 @@ impl From<Credentials> for proto::member::Credentials {
 }
 
 impl TryFrom<proto::create_request::El> for Endpoint {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(proto: proto::create_request::El) -> Result<Self, Self::Error> {
-        use proto::create_request::El as proto_el;
+    fn try_from(val: proto::create_request::El) -> Result<Self, Self::Error> {
+        use proto::create_request::El;
 
-        match proto {
-            proto_el::WebrtcPlay(play) => play.try_into().map(Self::WebRtcPlay),
-            proto_el::WebrtcPub(publish) => {
-                Ok(Self::WebRtcPublish(publish.into()))
-            }
-            proto_el::Room(proto::Room { id, .. })
-            | proto_el::Member(proto::Member { id, .. }) => {
-                Err(TryFromProtobufError::ExpectedOtherElement("Endpoint", id))
+        match val {
+            El::WebrtcPlay(play) => play.try_into().map(Self::WebRtcPlay),
+            El::WebrtcPub(publish) => Ok(Self::WebRtcPublish(publish.into())),
+            El::Room(proto::Room { id, .. })
+            | El::Member(proto::Member { id, .. }) => {
+                Err(ProtobufError::ExpectedElement("Endpoint", id.into()))
             }
         }
     }
 }
 
 impl TryFrom<proto::member::element::El> for Endpoint {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(
-        proto: proto::member::element::El,
-    ) -> Result<Self, Self::Error> {
+    fn try_from(val: proto::member::element::El) -> Result<Self, Self::Error> {
         use proto::member::element::El;
 
-        match proto {
+        match val {
             El::WebrtcPlay(e) => WebRtcPlay::try_from(e).map(Self::WebRtcPlay),
             El::WebrtcPub(e) => Ok(Self::WebRtcPublish(e.into())),
         }
@@ -504,21 +482,21 @@ impl TryFrom<proto::member::element::El> for Endpoint {
 }
 
 impl From<Endpoint> for proto::member::Element {
-    fn from(endpoint: Endpoint) -> Self {
-        use proto::member::element::El as proto_el;
+    fn from(val: Endpoint) -> Self {
+        use proto::member::element::El;
 
         Self {
-            el: Some(match endpoint {
-                Endpoint::WebRtcPlay(e) => proto_el::WebrtcPlay(e.into()),
-                Endpoint::WebRtcPublish(e) => proto_el::WebrtcPub(e.into()),
+            el: Some(match val {
+                Endpoint::WebRtcPlay(e) => El::WebrtcPlay(e.into()),
+                Endpoint::WebRtcPublish(e) => El::WebrtcPub(e.into()),
             }),
         }
     }
 }
 
 impl From<Endpoint> for proto::create_request::El {
-    fn from(endpoint: Endpoint) -> Self {
-        match endpoint {
+    fn from(val: Endpoint) -> Self {
+        match val {
             Endpoint::WebRtcPublish(publish) => Self::WebrtcPub(publish.into()),
             Endpoint::WebRtcPlay(play) => Self::WebrtcPlay(play.into()),
         }
@@ -526,8 +504,8 @@ impl From<Endpoint> for proto::create_request::El {
 }
 
 impl From<Endpoint> for proto::apply_request::El {
-    fn from(endpoint: Endpoint) -> Self {
-        match endpoint {
+    fn from(val: Endpoint) -> Self {
+        match val {
             Endpoint::WebRtcPublish(publish) => Self::WebrtcPub(publish.into()),
             Endpoint::WebRtcPlay(play) => Self::WebrtcPlay(play.into()),
         }
@@ -535,71 +513,71 @@ impl From<Endpoint> for proto::apply_request::El {
 }
 
 impl TryFrom<proto::WebRtcPlayEndpoint> for WebRtcPlay {
-    type Error = TryFromProtobufError;
+    type Error = ProtobufError;
 
-    fn try_from(value: proto::WebRtcPlayEndpoint) -> Result<Self, Self::Error> {
+    fn try_from(val: proto::WebRtcPlayEndpoint) -> Result<Self, Self::Error> {
         Ok(Self {
-            id: value.id.into(),
-            src: value.src.parse()?,
-            force_relay: value.force_relay,
+            id: val.id.into(),
+            src: val.src.parse()?,
+            force_relay: val.force_relay,
         })
     }
 }
 
 impl From<WebRtcPlay> for proto::WebRtcPlayEndpoint {
-    fn from(play: WebRtcPlay) -> Self {
+    fn from(val: WebRtcPlay) -> Self {
         Self {
-            id: play.id.into(),
-            src: play.src.to_string(),
+            id: val.id.into(),
+            src: val.src.to_string(),
             on_start: String::new(),
             on_stop: String::new(),
-            force_relay: play.force_relay,
+            force_relay: val.force_relay,
         }
     }
 }
 
 impl From<proto::WebRtcPublishEndpoint> for WebRtcPublish {
-    fn from(value: proto::WebRtcPublishEndpoint) -> Self {
+    fn from(val: proto::WebRtcPublishEndpoint) -> Self {
         use proto::web_rtc_publish_endpoint::P2p;
 
         Self {
-            id: value.id.into(),
-            p2p: P2p::from_i32(value.p2p).unwrap_or_default().into(),
-            audio_settings: value
+            id: val.id.into(),
+            p2p: P2p::from_i32(val.p2p).unwrap_or_default().into(),
+            audio_settings: val
                 .audio_settings
                 .map(AudioSettings::from)
                 .unwrap_or_default(),
-            video_settings: value
+            video_settings: val
                 .video_settings
                 .map(VideoSettings::from)
                 .unwrap_or_default(),
-            force_relay: value.force_relay,
+            force_relay: val.force_relay,
         }
     }
 }
 
 impl From<WebRtcPublish> for proto::WebRtcPublishEndpoint {
-    fn from(publish: WebRtcPublish) -> Self {
+    fn from(val: WebRtcPublish) -> Self {
         use proto::web_rtc_publish_endpoint::P2p;
 
         Self {
-            id: publish.id.into(),
-            p2p: P2p::from(publish.p2p).into(),
+            id: val.id.into(),
+            p2p: P2p::from(val.p2p).into(),
             on_start: String::new(),
             on_stop: String::new(),
-            force_relay: publish.force_relay,
-            audio_settings: Some(publish.audio_settings.into()),
-            video_settings: Some(publish.video_settings.into()),
+            force_relay: val.force_relay,
+            audio_settings: Some(val.audio_settings.into()),
+            video_settings: Some(val.video_settings.into()),
         }
     }
 }
 
 impl From<proto::web_rtc_publish_endpoint::AudioSettings> for AudioSettings {
-    fn from(from: proto::web_rtc_publish_endpoint::AudioSettings) -> Self {
+    fn from(val: proto::web_rtc_publish_endpoint::AudioSettings) -> Self {
         use proto::web_rtc_publish_endpoint::PublishPolicy;
 
         Self {
-            publish_policy: PublishPolicy::from_i32(from.publish_policy)
+            publish_policy: PublishPolicy::from_i32(val.publish_policy)
                 .unwrap_or_default()
                 .into(),
         }
@@ -617,11 +595,11 @@ impl From<AudioSettings> for proto::web_rtc_publish_endpoint::AudioSettings {
 }
 
 impl From<proto::web_rtc_publish_endpoint::VideoSettings> for VideoSettings {
-    fn from(from: proto::web_rtc_publish_endpoint::VideoSettings) -> Self {
+    fn from(val: proto::web_rtc_publish_endpoint::VideoSettings) -> Self {
         use proto::web_rtc_publish_endpoint::PublishPolicy;
 
         Self {
-            publish_policy: PublishPolicy::from_i32(from.publish_policy)
+            publish_policy: PublishPolicy::from_i32(val.publish_policy)
                 .unwrap_or_default()
                 .into(),
         }
@@ -639,10 +617,10 @@ impl From<VideoSettings> for proto::web_rtc_publish_endpoint::VideoSettings {
 }
 
 impl From<proto::web_rtc_publish_endpoint::PublishPolicy> for Policy {
-    fn from(from: proto::web_rtc_publish_endpoint::PublishPolicy) -> Self {
+    fn from(val: proto::web_rtc_publish_endpoint::PublishPolicy) -> Self {
         use proto::web_rtc_publish_endpoint::PublishPolicy as Proto;
 
-        match from {
+        match val {
             Proto::Optional => Self::Optional,
             Proto::Required => Self::Required,
             Proto::Disabled => Self::Disabled,
@@ -651,8 +629,8 @@ impl From<proto::web_rtc_publish_endpoint::PublishPolicy> for Policy {
 }
 
 impl From<Policy> for proto::web_rtc_publish_endpoint::PublishPolicy {
-    fn from(from: Policy) -> Self {
-        match from {
+    fn from(val: Policy) -> Self {
+        match val {
             Policy::Optional => Self::Optional,
             Policy::Required => Self::Required,
             Policy::Disabled => Self::Disabled,
@@ -661,10 +639,10 @@ impl From<Policy> for proto::web_rtc_publish_endpoint::PublishPolicy {
 }
 
 impl From<proto::web_rtc_publish_endpoint::P2p> for P2pMode {
-    fn from(value: proto::web_rtc_publish_endpoint::P2p) -> Self {
+    fn from(val: proto::web_rtc_publish_endpoint::P2p) -> Self {
         use proto::web_rtc_publish_endpoint::P2p;
 
-        match value {
+        match val {
             P2p::Always => Self::Always,
             P2p::IfPossible => Self::IfPossible,
             P2p::Never => Self::Never,
@@ -673,8 +651,8 @@ impl From<proto::web_rtc_publish_endpoint::P2p> for P2pMode {
 }
 
 impl From<P2pMode> for proto::web_rtc_publish_endpoint::P2p {
-    fn from(value: P2pMode) -> Self {
-        match value {
+    fn from(val: P2pMode) -> Self {
+        match val {
             P2pMode::Always => Self::Always,
             P2pMode::IfPossible => Self::IfPossible,
             P2pMode::Never => Self::Never,
@@ -683,25 +661,25 @@ impl From<P2pMode> for proto::web_rtc_publish_endpoint::P2p {
 }
 
 impl From<proto::Ping> for Ping {
-    fn from(v: proto::Ping) -> Self {
-        Self(v.nonce)
+    fn from(val: proto::Ping) -> Self {
+        Self(val.nonce)
     }
 }
 
 impl From<Ping> for proto::Ping {
-    fn from(v: Ping) -> Self {
-        Self { nonce: v.0 }
+    fn from(val: Ping) -> Self {
+        Self { nonce: val.0 }
     }
 }
 
 impl From<proto::Pong> for Pong {
-    fn from(v: proto::Pong) -> Self {
-        Self(v.nonce)
+    fn from(val: proto::Pong) -> Self {
+        Self(val.nonce)
     }
 }
 
 impl From<Pong> for proto::Pong {
-    fn from(v: Pong) -> Self {
-        Self { nonce: v.0 }
+    fn from(val: Pong) -> Self {
+        Self { nonce: val.0 }
     }
 }