+++ proto/control-api/src/grpc/mod.rs
@@ -1,12 +1,15 @@
-//! [gRPC]-based [Control API] for [Medea].
+//! [gRPC]-based [Control API] implementation.
 //!
 //! [gRPC]: https://grpc.io
-//! [Medea]: https://github.com/instrumentisto/medea
 //! [Control API]: https://tinyurl.com/yxsqplq7
+
+// TODO: Remove once annoying false positive is fixed:
+//       https://github.com/rust-lang/rust-clippy/issues/6902
+#![allow(clippy::use_self)]
 
 #[cfg(feature = "client")]
 mod client;
-mod conversions;
+mod convert;
 #[cfg(feature = "server")]
 mod server;
 
@@ -53,10 +56,13 @@ pub mod api;
 #[rustfmt::skip]
 pub mod callback;
 
+use std::str::FromStr;
+
+use derive_more::{Display, Error, From, Into};
+use url::Url;
+
 #[doc(inline)]
-pub use self::conversions::{
-    CallbackUrl, CallbackUrlParseError, TryFromProtobufError,
-};
+pub use self::convert::ProtobufError;
 #[cfg(feature = "client")]
 #[doc(inline)]
 pub use self::{
@@ -70,3 +76,53 @@ pub use self::{
     api::control_api_server::ControlApiServer,
     callback::callback_server::CallbackServer,
 };
+
+/// URL representing a [gRPC] callback implementing [`CallbackApi`].
+///
+/// [`CallbackApi`]: crate::CallbackApi
+/// [gRPC]: https://grpc.io
+#[derive(Clone, Debug, Display, Eq, Hash, Into, PartialEq)]
+#[display(fmt = "grpc://{}", _0)]
+#[into(owned(types(String)))]
+pub struct CallbackUrl(Url);
+
+impl CallbackUrl {
+    /// Converts this [`CallbackUrl`] into the one with `http://` scheme.
+    #[must_use]
+    pub fn to_http(&self) -> Url {
+        let mut url = self.0.clone();
+        url.set_scheme("http").unwrap_or_else(|_| unreachable!());
+        url
+    }
+}
+
+impl FromStr for CallbackUrl {
+    type Err = CallbackUrlParseError;
+
+    fn from_str(value: &str) -> Result<Self, Self::Err> {
+        let url = Url::parse(value)?;
+        if url.scheme() != "grpc" {
+            return Err(CallbackUrlParseError::WrongScheme);
+        }
+        if url.host().is_none() {
+            return Err(CallbackUrlParseError::MissingHost);
+        }
+        Ok(Self(url))
+    }
+}
+
+/// Error of parsing a [`CallbackUrl`].
+#[derive(Clone, Copy, Debug, Display, Error, From)]
+pub enum CallbackUrlParseError {
+    /// Error parsing the provided [`Url`].
+    #[display(fmt = "Invalid URL: {}", _0)]
+    UrlParseErr(url::ParseError),
+
+    /// [`Url`] is missing host.
+    #[display(fmt = "Missing host")]
+    MissingHost,
+
+    /// [`Url`] contains unsupported scheme.
+    #[display(fmt = "Only `grpc://` scheme is allowed")]
+    WrongScheme,
+}